+++ CoffeeShop.Web/Api/ApplicationRoleController.cs
@@ -119,7 +119,7 @@ namespace CoffeeShop.Web.Api
                 var appRole = _appRoleService.GetDetail(applicationRoleViewModel.Id);
                 try
                 {
-                    appRole.UpdateApplicationRole(applicationRoleViewModel, "update");
+                    appRole.UpdateApplicationRole(applicationRoleViewModel, "Update");
                     _appRoleService.Update(appRole);
                     _appRoleService.SaveChanges();
                     return request.CreateResponse(HttpStatusCode.OK, appRole);
@@ -139,6 +139,10 @@ namespace CoffeeShop.Web.Api
         [Route("Delete")]
         public HttpResponseMessage Delete(HttpRequestMessage request, string id)
         {
+            var deleteRole = _appRoleService.GetByStringId(id);
+            if(deleteRole == null || deleteRole.IsSystemProtected)
+                return request.CreateResponse(HttpStatusCode.NotAcceptable, $"You cannot delete this role \"{deleteRole.Name}\" because it is protected by the system, Contact authorized person for more details");
+
             _appRoleService.Delete(id);
             _appRoleService.SaveChanges();
             return request.CreateResponse(HttpStatusCode.OK, id);
@@ -160,6 +164,10 @@ namespace CoffeeShop.Web.Api
                     var listItem = new JavaScriptSerializer().Deserialize<List<string>>(ids);
                     foreach (var item in listItem)
                     {
+                        var deleteRole = _appRoleService.GetByStringId(item);
+                        if (deleteRole == null || deleteRole.IsSystemProtected)
+                            return request.CreateResponse(HttpStatusCode.NotAcceptable, $"You cannot delete this role \"{deleteRole.Name}\" because it is protected by the system, Contact authorized person for more details");
+
                         _appRoleService.Delete(item);
                     }
